pyproject.toml
[tool.poetry]
name = "real-estate-saas-core"
version = "0.2.0"
description = "Production-oriented async real estate scraping platform."
authors = ["Parser Team <team@example.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = { extras = ["standard"], version = "^0.32.0" }
httpx = "^0.28.0"
lxml = "^6.0.3"
playwright = "^1.55.0"
aiomysql = "^0.2.0"
mysql-connector-python = "^9.0.0"
redis = "^5.2.0"
scrapling = { extras = ["fetchers"], version = "^0.4.6" }
twilio = "^9.3.0"
google-cloud-speech = "^2.30.0"
google-cloud-texttospeech = "^2.23.0"
pydantic = "^2.10.0"
pydantic-settings = "^2.6.0"
python-dotenv = "^1.0.1"
python-multipart = "^0.0.19"
streamlit = "^1.40.0"
streamlit-authenticator = "^0.4.1"
pyyaml = "^6.0.2"
passlib = "^1.7.4"
pyjwt = "^2.9.0"
email-validator = "^2.2.0"
sentry-sdk = { extras = ["aiohttp", "httpx"], version = "^2.19.2" }
structlog = "^24.4.0"
pandas = "^2.2.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.7.4"
mypy = "^1.13.0"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "-p no:cacheprovider"
testpaths = ["tests"]
norecursedirs = ["pytest-cache-files-*", ".git", "venv"]

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"
